Change how the client gives player items to use new `get_item_array`
This method uses an asm patch to add an array that the game checks every frame. If you place an item ID in the array, the game will give the player that item.
So, the client only needs to place the item IDs of the item it wants to give the player in that array in RAM.
This greatly simplifies the client code because it no longer has to manually set bits in memory or keep track of various addresses.