Refactor ER code
This update uses the code that the base randomizer uses for ER, with some modifications for AP.
This should resolve some generation issues that the randomizer has. If there is an issue, it is probably in the base randomizer too.
One of the main thing this addresses is to not put a banned and required dungeon on the same island.