NumRequiredBosses Tooltip
The existing tooltip leads players to believe that there is either nothing in these locations at all, or that the items present have no value.  However, numerous other worlds either misuse the filler tag, or dump helpful but not critical items into it.  As a result, many players will claim they have nothing to do in their world when some of these locations will stay have meaningful value to multiworld participants.  This change clarifies the filler class and emphasizes the lack of progression, to at least minimize the number of people who might believe the locations simply do not exist.